+++ benches/benches/block_target_gas_set/mod.rs
@@ -1,3 +1,9 @@
 pub mod alu;
 
 pub mod crypto;
+
+pub mod flow;
+
+pub mod contract;
+
+pub mod memory;