+++ benches/benches/vm_set/crypto.rs
@@ -73,7 +73,7 @@ pub fn run(c: &mut Criterion) {
             ),
     );
 
-    let linear = super::utils::generate_linear_costs();
+    let linear = super::utils::arb_dependent_cost_values();
 
     let mut bench_k256 = c.benchmark_group("k256");
     for i in &linear {