+++ crates/fuel-core/src/p2p_test_helpers.rs
@@ -14,6 +14,8 @@ use crate::{
 use fuel_core_p2p::{
     codecs::postcard::PostcardCodec,
     network_service::FuelP2PService,
+    p2p_service::FuelP2PEvent,
+    service::to_message_acceptance,
 };
 use fuel_core_poa::{
     ports::BlockImporter,
@@ -43,6 +45,7 @@ use fuel_core_types::{
         ChainId,
     },
     secrecy::Secret,
+    services::p2p::GossipsubMessageAcceptance,
 };
 use futures::StreamExt;
 use itertools::Itertools;
@@ -62,6 +65,12 @@ use std::{
 };
 use tokio::sync::broadcast;
 
+#[derive(Copy, Clone)]
+pub enum BootstrapType {
+    BootstrapNodes,
+    ReservedNodes,
+}
+
 #[derive(Clone)]
 /// Setup for a producer node
 pub struct ProducerSetup {
@@ -71,6 +80,10 @@ pub struct ProducerSetup {
     pub secret: SecretKey,
     /// Number of test transactions to create for this producer.
     pub num_test_txs: usize,
+    /// Enable full utxo stateful validation.
+    pub utxo_validation: bool,
+    /// Indicates the type of initial connections.
+    pub bootstrap_type: BootstrapType,
 }
 
 #[derive(Clone)]
@@ -80,6 +93,10 @@ pub struct ValidatorSetup {
     pub name: String,
     /// Public key of the producer to sync from.
     pub pub_key: Address,
+    /// Enable full utxo stateful validation.
+    pub utxo_validation: bool,
+    /// Indicates the type of initial connections.
+    pub bootstrap_type: BootstrapType,
 }
 
 #[derive(Clone)]
@@ -126,7 +143,20 @@ impl Bootstrap {
                         assert!(result.is_ok());
                         break;
                     }
-                    _ = bootstrap.next_event() => {}
+                    event = bootstrap.next_event() => {
+                        // The bootstrap node only forwards data without validating it.
+                        if let Some(FuelP2PEvent::GossipsubMessage {
+                            peer_id,
+                            message_id,
+                            ..
+                        }) = event {
+                            bootstrap.report_message_validation_result(
+                                &message_id,
+                                peer_id,
+                                to_message_acceptance(&GossipsubMessageAcceptance::Accept)
+                            )
+                        }
+                    }
                 }
             }
         });
@@ -256,9 +286,27 @@ pub async fn make_nodes(
 
         let mut test_txs = Vec::with_capacity(0);
         node_config.block_production = Trigger::Instant;
-        node_config.p2p.as_mut().unwrap().bootstrap_nodes = boots.clone();
 
-        if let Some((ProducerSetup { secret, .. }, txs)) = s {
+        if let Some((
+            ProducerSetup {
+                secret,
+                utxo_validation,
+                bootstrap_type,
+                ..
+            },
+            txs,
+        )) = s
+        {
+            match bootstrap_type {
+                BootstrapType::BootstrapNodes => {
+                    node_config.p2p.as_mut().unwrap().bootstrap_nodes = boots.clone();
+                }
+                BootstrapType::ReservedNodes => {
+                    node_config.p2p.as_mut().unwrap().reserved_nodes = boots.clone();
+                }
+            }
+
+            node_config.utxo_validation = utxo_validation;
             let pub_key = secret.public_key();
             match &mut node_config.chain_conf.consensus {
                 crate::chain_config::ConsensusConfig::PoA { signing_key } => {
@@ -286,9 +334,24 @@ pub async fn make_nodes(
             chain_config.clone(),
         );
         node_config.block_production = Trigger::Never;
-        node_config.p2p.as_mut().unwrap().bootstrap_nodes = boots.clone();
 
-        if let Some(ValidatorSetup { pub_key, .. }) = s {
+        if let Some(ValidatorSetup {
+            pub_key,
+            utxo_validation,
+            bootstrap_type,
+            ..
+        }) = s
+        {
+            node_config.utxo_validation = utxo_validation;
+
+            match bootstrap_type {
+                BootstrapType::BootstrapNodes => {
+                    node_config.p2p.as_mut().unwrap().bootstrap_nodes = boots.clone();
+                }
+                BootstrapType::ReservedNodes => {
+                    node_config.p2p.as_mut().unwrap().reserved_nodes = boots.clone();
+                }
+            }
             match &mut node_config.chain_conf.consensus {
                 crate::chain_config::ConsensusConfig::PoA { signing_key } => {
                     *signing_key = pub_key;
@@ -343,6 +406,11 @@ fn extract_p2p_config(node_config: &Config) -> fuel_core_p2p::config::Config {
 }
 
 impl Node {
+    /// Returns the vector of valid transactions for pre-initialized state.
+    pub fn test_transactions(&self) -> &Vec<Transaction> {
+        &self.test_txs
+    }
+
     /// Waits for `number_of_blocks` and each block should be `is_local`
     pub async fn wait_for_blocks(&self, number_of_blocks: usize, is_local: bool) {
         let mut stream = self
@@ -453,6 +521,8 @@ impl ProducerSetup {
             name: Default::default(),
             secret,
             num_test_txs: Default::default(),
+            utxo_validation: true,
+            bootstrap_type: BootstrapType::BootstrapNodes,
         }
     }
 
@@ -469,6 +539,20 @@ impl ProducerSetup {
             ..self
         }
     }
+
+    pub fn utxo_validation(self, utxo_validation: bool) -> Self {
+        Self {
+            utxo_validation,
+            ..self
+        }
+    }
+
+    pub fn bootstrap_type(self, bootstrap_type: BootstrapType) -> Self {
+        Self {
+            bootstrap_type,
+            ..self
+        }
+    }
 }
 
 impl ValidatorSetup {
@@ -476,12 +560,28 @@ impl ValidatorSetup {
         Self {
             pub_key,
             name: Default::default(),
+            utxo_validation: true,
+            bootstrap_type: BootstrapType::BootstrapNodes,
         }
     }
 
     pub fn with_name(self, name: impl Into<String>) -> Self {
         Self {
             name: name.into(),
+            ..self
+        }
+    }
+
+    pub fn utxo_validation(self, utxo_validation: bool) -> Self {
+        Self {
+            utxo_validation,
+            ..self
+        }
+    }
+
+    pub fn bootstrap_type(self, bootstrap_type: BootstrapType) -> Self {
+        Self {
+            bootstrap_type,
             ..self
         }
     }